+++ storage/src/raft_log_store/file.rs
@@ -1,36 +1,40 @@
-use std::io::SeekFrom;
 use std::path::Path;
+use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::Arc;
 
-use tokio::fs;
-use tokio::io::{AsyncReadExt, AsyncSeekExt, AsyncWriteExt};
-use tokio::sync::Mutex;
-use tracing::{trace_span, Instrument};
+use tokio::sync::oneshot;
 
 use crate::error::Result;
+
+pub async fn sync_dir(dir: impl AsRef<Path>) -> Result<()> {
+    let path = dir.as_ref().to_path_buf();
+
+    let (tx, rx) = oneshot::channel();
+    tokio::task::spawn_blocking(move || {
+        let dir = match std::fs::File::open(path) {
+            Ok(dir) => dir,
+            Err(e) => {
+                tx.send(Err(e)).unwrap();
+                return;
+            }
+        };
+        tx.send(dir.sync_all()).unwrap();
+    });
+
+    rx.await.unwrap()?;
+
+    Ok(())
+}
 
 fn filename(id: u64) -> String {
     format!("{:08}", id)
 }
 
-pub async fn sync_dir(dir: impl AsRef<Path>) -> Result<()> {
-    fs::File::open(dir.as_ref())
-        .await?
-        .sync_all()
-        .instrument(trace_span!("sync_all"))
-        .await?;
-    Ok(())
-}
-
-struct ActiveFileCore {
-    file: fs::File,
-    offset: usize,
-}
-
 #[derive(Clone)]
 pub struct ActiveFile {
     id: u64,
-    core: Arc<Mutex<ActiveFileCore>>,
+    file: Arc<std::fs::File>,
+    len: Arc<AtomicUsize>,
 }
 
 impl std::fmt::Debug for ActiveFile {
@@ -42,15 +46,24 @@ impl std::fmt::Debug for ActiveFile {
 impl ActiveFile {
     #[tracing::instrument(level = "trace", skip(dir), err)]
     pub async fn open(dir: impl AsRef<Path>, id: u64) -> Result<Self> {
-        let mut options = fs::OpenOptions::new();
-        options.create(true);
-        options.read(true);
-        options.append(true);
-        let file = options.open(dir.as_ref().join(filename(id))).await?;
+        let (tx, rx) = oneshot::channel();
+        let path = dir.as_ref().join(filename(id));
+
+        tokio::task::spawn_blocking(move || {
+            let mut options = std::fs::OpenOptions::new();
+            options.create(true);
+            options.append(true);
+            options.read(true);
+            let result = options.open(path);
+            tx.send(result).unwrap();
+        });
+
+        let file = rx.await.unwrap()?;
 
         Ok(Self {
             id,
-            core: Arc::new(Mutex::new(ActiveFileCore { file, offset: 0 })),
+            file: Arc::new(file),
+            len: Arc::new(AtomicUsize::new(0)),
         })
     }
 
@@ -60,71 +73,105 @@ impl ActiveFile {
     }
 
     #[tracing::instrument(level = "trace", ret)]
-    pub async fn offset(&self) -> u64 {
-        self.core.lock().await.offset as u64
+    pub fn is_empty(&self) -> bool {
+        self.len() == 0
     }
 
-    #[tracing::instrument(level = "trace", err)]
-    pub async fn append(&self, buf: &[u8]) -> Result<(u64, usize)> {
-        let mut core = self.core.lock().await;
-        let offset = core.offset as u64;
-        let len = buf.len();
-        core.file
-            .write_all(buf)
-            .instrument(trace_span!("write_all"))
-            .await?;
-        core.offset += len;
-        Ok((offset, len))
+    #[tracing::instrument(level = "trace", ret)]
+    pub fn len(&self) -> usize {
+        self.len.load(Ordering::Acquire)
     }
 
     #[tracing::instrument(level = "trace", err)]
     pub async fn read(&self, offset: u64, len: usize) -> Result<Vec<u8>> {
-        let mut core = self.core.lock().await;
-        core.file.seek(SeekFrom::Start(offset)).await?;
-        let mut buf = vec![0; len];
-        core.file.read_exact(&mut buf).await?;
+        let (tx, rx) = oneshot::channel();
+        let f = self.file.clone();
+
+        tokio::task::spawn_blocking(move || {
+            use std::os::unix::prelude::FileExt;
+            let mut buf = vec![0; len];
+            let result = f.read_exact_at(&mut buf, offset).map(|_| buf);
+            tx.send(result).unwrap();
+        });
+
+        let buf = rx.await.unwrap()?;
+
         Ok(buf)
     }
 
+    // TODO: Replace `buf: Vec<u8>` with `buf: Bytes` or `buf: &[u8]` (but 'static is needed).
+    #[tracing::instrument(level = "trace", err)]
+    pub async fn append(&self, buf: Vec<u8>) -> Result<()> {
+        let (tx, rx) = oneshot::channel();
+        let f = self.file.clone();
+        let buf_len = buf.len();
+
+        tokio::task::spawn_blocking(move || {
+            use std::os::unix::prelude::FileExt;
+            let result = f.write_all_at(&buf, 0);
+            tx.send(result).unwrap();
+        });
+
+        rx.await.unwrap()?;
+        self.len.fetch_add(buf_len, Ordering::Release);
+
+        Ok(())
+    }
+
+    /// # Safety
+    /// Concurrent `flush` call leads to panics.
     #[tracing::instrument(level = "trace", err)]
     pub async fn flush(&self) -> Result<()> {
-        self.core.lock().await.file.flush().await?;
+        let (tx, rx) = oneshot::channel();
+        let mut f = self.file.clone();
+
+        tokio::task::spawn_blocking(move || {
+            use std::io::Write;
+            let fmut = Arc::get_mut(&mut f).unwrap();
+            let result = fmut.flush();
+            tx.send(result).unwrap();
+        });
+
+        rx.await.unwrap()?;
+
         Ok(())
     }
 
     #[tracing::instrument(level = "trace", err)]
     pub async fn sync_data(&self) -> Result<()> {
-        self.core
-            .lock()
-            .await
-            .file
-            .sync_data()
-            .instrument(trace_span!("sync_data"))
-            .await?;
+        let (tx, rx) = oneshot::channel();
+        let f = self.file.clone();
+
+        tokio::task::spawn_blocking(move || {
+            let result = f.sync_data();
+            tx.send(result).unwrap();
+        });
+
+        rx.await.unwrap()?;
+
         Ok(())
     }
 
     #[tracing::instrument(level = "trace", err)]
     pub async fn sync_all(&self) -> Result<()> {
-        self.core
-            .lock()
-            .await
-            .file
-            .sync_all()
-            .instrument(trace_span!("sync_all"))
-            .await?;
+        let (tx, rx) = oneshot::channel();
+        let f = self.file.clone();
+
+        tokio::task::spawn_blocking(move || {
+            let result = f.sync_all();
+            tx.send(result).unwrap();
+        });
+
+        rx.await.unwrap()?;
+
         Ok(())
     }
-}
-
-struct FrozenFileCore {
-    file: fs::File,
 }
 
 #[derive(Clone)]
 pub struct FrozenFile {
     id: u64,
-    core: Arc<Mutex<FrozenFileCore>>,
+    file: Arc<std::fs::File>,
 }
 
 impl std::fmt::Debug for FrozenFile {
@@ -136,13 +183,21 @@ impl std::fmt::Debug for FrozenFile {
 impl FrozenFile {
     #[tracing::instrument(level = "trace", skip(dir), err)]
     pub async fn open(dir: impl AsRef<Path>, id: u64) -> Result<Self> {
-        let mut options = fs::OpenOptions::new();
-        options.read(true);
-        let file = options.open(dir.as_ref().join(filename(id))).await?;
+        let (tx, rx) = oneshot::channel();
+        let path = dir.as_ref().join(filename(id));
+
+        tokio::task::spawn_blocking(move || {
+            let mut options = std::fs::OpenOptions::new();
+            options.read(true);
+            let result = options.open(path);
+            tx.send(result).unwrap();
+        });
+
+        let file = rx.await.unwrap()?;
 
         Ok(Self {
             id,
-            core: Arc::new(Mutex::new(FrozenFileCore { file })),
+            file: Arc::new(file),
         })
     }
 
@@ -153,19 +208,55 @@ impl FrozenFile {
 
     #[tracing::instrument(level = "trace", err)]
     pub async fn read(&self, offset: u64, len: usize) -> Result<Vec<u8>> {
-        let mut core = self.core.lock().await;
-        core.file.seek(SeekFrom::Start(offset)).await?;
-        let mut buf = vec![0; len];
-        core.file.read_exact(&mut buf).await?;
+        let (tx, rx) = oneshot::channel();
+        let f = self.file.clone();
+
+        tokio::task::spawn_blocking(move || {
+            use std::os::unix::prelude::FileExt;
+            let mut buf = vec![0; len];
+            let result = f.read_exact_at(&mut buf, offset).map(|_| buf);
+            tx.send(result).unwrap();
+        });
+
+        let buf = rx.await.unwrap()?;
+
         Ok(buf)
     }
 
     #[tracing::instrument(level = "trace", err)]
     pub async fn read_all(&self) -> Result<Vec<u8>> {
-        let mut core = self.core.lock().await;
-        core.file.seek(SeekFrom::Start(0)).await?;
-        let mut buf = Vec::new();
-        core.file.read_to_end(&mut buf).await?;
+        let (tx, rx) = oneshot::channel();
+        let f = self.file.clone();
+
+        tokio::task::spawn_blocking(move || {
+            use std::os::unix::prelude::FileExt;
+            let len = match f.metadata() {
+                Ok(metadata) => metadata.len() as usize,
+                Err(e) => {
+                    tx.send(Err(e)).unwrap();
+                    return;
+                }
+            };
+            let mut buf = vec![0; len];
+            let result = f.read_exact_at(&mut buf, 0).map(|_| buf);
+            tx.send(result).unwrap();
+        });
+
+        let buf = rx.await.unwrap()?;
+
         Ok(buf)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn is_send_sync_clone<T: Send + Sync + Clone + 'static>() {}
+
+    #[test]
+    fn ensure_send_sync_clone() {
+        is_send_sync_clone::<ActiveFile>();
+        is_send_sync_clone::<FrozenFile>();
     }
 }
